cloud_functions/interest_rates/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.11.4
requests>=2.28.0
